Bouncing balls the game :game_die:  :video_game:

+++ Multiplayer Game/ModuleGameObject.h
@@ -7,7 +7,7 @@ struct GameObject
 	// Transform component
 	vec2 position = vec2{ 0.0f, 0.0f };
 	vec2 size = vec2{ 0.0f, 0.0f }; // NOTE(jesus): If equals 0, it takes the size of the texture
-	float angle = 0.0f;
+	int angle = 0;
 
 	// Render component
 	Sprite *sprite = nullptr;
